+++ Server/Single Thread ASIO/source/Protocol.h
@@ -109,17 +109,17 @@ struct CS_MAP_LOADED_PACKET {		// 클라이언트 map 로드 완료
 struct CS_MOVE_PACKET {				// 플레이어 움직임
 	unsigned char	size;
 	char			type;
-	float			rx, ry, rz;
-	float			x, y, z;
-	float			speed;
+	double			rx, ry, rz;
+	double			x, y, z;
+	double			speed;
 	bool			jump;
 };
 
 struct CS_ATTACK_PACKET {			// 플레이어 때림 애니메이션
 	unsigned char	size;
 	char			type;
-	float			rx, ry, rz;
-	float			x, y, z;
+	double			rx, ry, rz;
+	double			x, y, z;
 };
 
 struct CS_PICKUP_FUSE_PACKET {		// 플레이어 아이템 얻음
@@ -168,8 +168,8 @@ struct CS_IDLE_STATE_PACKET {
 struct CS_CANNON_FIRE_PACKET {
 	unsigned char	size;
 	char			type;
-	float			x, y, z;
-	float			yaw, pitch;
+	double			x, y, z;
+	double			yaw, pitch;
 };
 
 // ====================================== 서버 -> 클라 패킷 ==========================================
@@ -211,7 +211,7 @@ struct SC_ADD_PLAYER_PACKET {		// 플레이어 추가
 	char			type;
 	int				id;
 	int				_hp;
-	float			x, y, z;
+	double			x, y, z;
 	char			role[PROTOCOL_NAME_SIZE];
 	int				charactorNum;				// 1~5 생존자, 6~7 살인마
 };
@@ -220,9 +220,9 @@ struct SC_MOVE_PLAYER_PACKET {		// 플레이어 움직임
 	unsigned char	size;
 	char			type;
 	int				id;
-	float			x, y, z;
-	float			rx, ry, rz;
-	float			speed;
+	double			x, y, z;
+	double			rx, ry, rz;
+	double			speed;
 	bool			jump;
 };
 
@@ -276,7 +276,7 @@ struct SC_OPENING_ITEM_BOX_PACKET {
 	char			type;
 	int				id;
 	int				index;
-	float			progress;
+	double			progress;
 };
 
 struct SC_STOP_OPENING_PACKET {
@@ -285,7 +285,7 @@ struct SC_STOP_OPENING_PACKET {
 	int				id;				
 	int				item;			
 	int				index;			
-	float			progress;		
+	double			progress;		
 };
 
 struct SC_OPENING_FUSE_BOX_PACKET {
@@ -293,7 +293,7 @@ struct SC_OPENING_FUSE_BOX_PACKET {
 	char			type;
 	int				id;
 	int				index;
-	float			progress;
+	double			progress;
 };
